requirements.txt
confluent-kafka
snowflake-connector-python
python-dotenv
pandas
numpy